+++ composeApp/src/commonMain/kotlin/feature/core/domain/model/Expense.kt
@@ -9,11 +9,11 @@ data class Expense(
     val id: ObjectId = ObjectId(),
     val idString: String = id.toHexString(),
     val categoryId: Int,
-    val typeId: Int,
+    val typeId: Long,
     val description: String,
     val isIncome: Boolean=false,
     val cost: Long = 0,
     val timestamp: Long = 0
 ){
-    constructor(): this(categoryId = 0 , typeId = 0 , description = "")
+    constructor(): this(categoryId = 0 , typeId = 0L , description = "")
 }
